requirements.txt
sentencepiece
transformers
torch
flask
flask_cors
starlette
aiofiles
python-multipart
Pillow